requirements.txt
Django>=5.0,<5.1
gunicorn>=21.2
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.0
stripe>=7.0
whitenoise>=6.6
requests>=2.30
requests-oauthlib>=1.3
django-allauth>=0.53
